fix(accident): accId를 openapiId로 변경 (#15)
- AccidentRepository, AccidentService에서 accId를 openapiId로 변경이 되어 있지 않아 변경해줌
- AccidentOpenAPI에서 돌발 아이디를 Long openapiId로 저장하는 부분에서 Integer형을 Long형으로 바로 형변환할 수 없는 문제 해결

+++ src/main/java/com/capstone/bbiyong/openapi/AccidentOpenAPI.java
@@ -97,7 +97,7 @@ public class AccidentOpenAPI implements OpenAPI {
         for (int i = 0; i < jsonArray.length(); i++) {
             jsonObject = (JSONObject) jsonArray.get(i);
 
-            Long openapiId = (Long) jsonObject.get("acc_id"); /*돌발 아이디*/
+            Long openapiId = Long.valueOf((Integer) jsonObject.get("acc_id")); /*돌발 아이디*/
             Integer intStartDate = (Integer) jsonObject.get("occr_date"); /*발생 일자*/
             int intStartTime = jsonObject.getInt("occr_time"); /*발생 시각*/
             Integer intEndDate = (Integer) jsonObject.get("exp_clr_date"); /*종료 일자*/
